requirements.txt
azure-ai-documentintelligence==1.0.0b4
azure-core==1.31.0
azure-identity==1.18.0
azure-storage-blob==12.23.1
ipycanvas==0.13.3
ipykernel==6.29.5
jupyter-bbox-widget==0.6.0
marker-pdf==0.2.17
matplotlib==3.9.2
notebook==7.2.2
numpy
openai==1.51.0
opencv-python==4.10.0.84
openpyxl==3.1.5
pandas==2.2.3
pdf2image==1.17.0
pydantic==2.9.2
PyMuPDF==1.24.11
pytesseract==0.3.13
python-dotenv==1.0.1
seaborn==0.13.2
surya-ocr==0.5.0
tiktoken==0.8.0
transformers==4.45.1